Wrapped the rolodex view in application view and successfully
