+++ src/app.js
@@ -8,27 +8,27 @@ import Rolodex from 'app/collections/rolodex';
 
 
 
-var hardCodedContacts = [
-  { name: "Small Fry", email: "smallfry@condiments.org", phone: "303-555-1000" },
-  { name: "Wee Chip", email: "weechip@condiments.org", phone: "44 7700 900104"},
-  { name: "Oui Jean", email: "ouijean@condiments.org", phone: "303-555-2000" },
-  { name: "Baby-Q", email: "bbq@condiments.org", phone: "n/a, too young" }
-];
+// var hardCodedContacts = [
+//   { name: "Small Fry", email: "smallfry@condiments.org", phone: "303-555-1000" },
+//   { name: "Wee Chip", email: "weechip@condiments.org", phone: "44 7700 900104"},
+//   { name: "Oui Jean", email: "ouijean@condiments.org", phone: "303-555-2000" },
+//   { name: "Baby-Q", email: "bbq@condiments.org", phone: "n/a, too young" }
+// ];
 
 
 $(document).ready(function() {
   var application = new Application();
-  var appView = new Application({
+  var appView = new ApplicationView({
     el: '#application',
     model: application
   });
 
-  var contactsList = new Rolodex(hardCodedContacts);
-  var rolodexView = new RolodexView({
-    el: '#contact-cards',
-    model: contactsList
-  });
+  // var contactsList = new Rolodex(hardCodedContacts);
+  // var rolodexView = new RolodexView({
+  //   el: '#contact-cards',
+  //   model: contactsList
+  // });
 
-  // appView.render();
-  rolodexView.render();
+  appView.render();
+  // rolodexView.render();
 });
